DnD_project/requirements.txt
astroid==2.3.3
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.1.1
django-extensions==2.2.5
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==20.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
